ints to floats

+++ core/src/model/collectable/AbstractBuff.java
@@ -2,22 +2,21 @@ package model.collectable;
 
 import com.badlogic.gdx.math.Vector2;
 
-import model.GameObjectInterface;
 import view.CustomSprite;
 
 public abstract class AbstractBuff extends AbstractCollectable {
 
-    public AbstractBuff(Vector2 position, int width, int height) {
+    public AbstractBuff(Vector2 position, float width, float height) {
 		super(position, width, height);
 	}
 
 	@Override
-    public int getX() {
+    public float getX() {
         return super.getX();
     }
 
     @Override
-    public int getY() {
+    public float getY() {
         return super.getY();
     }
